Edit transaction
