+++ app/src/transactions/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 transactions_blueprint = Blueprint('transactions_blueprint', __name__, url_prefix='')
 
-from .presentation import routes
+from .presentation import crud_transaction_routes, load_transactions_file_routes
