+++ app/templates/transactions/edit_transaction.html
@@ -37,11 +37,11 @@
             {% endif %}
         </div>
         <div class="mb-3">
-            {{ form.category.label(class="form-label") }}
-            {{ form.category(class="form-select") }}
-            {% if form.category.errors %}
+            {{ form.category_id.label(class="form-label") }}
+            {{ form.category_id(class="form-select") }}
+            {% if form.category_id.errors %}
                 <div class="invalid-feedback">
-                    {% for error in form.category.errors %}
+                    {% for error in form.category_id.errors %}
                         <span>{{ error }}</span><br>
                     {% endfor %}
                 </div>
